btc_predictor/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
hydra-core==1.1.0
pandas==1.2.4
pmdarima==1.8.2
python-dotenv==0.18.0
scikit-learn==0.24.2
seaborn==0.11.1
tensorflow==2.5.0
urllib3==1.25.11
wandb==0.10.33
xgboost==1.4.2
